parser: support variable variable property fetch and method calls
